+++ commands/utility/etcGame.js
@@ -1,6 +1,7 @@
 const { SlashCommandBuilder } = require('discord.js');
 const partyRecruitmentsDao = require('../../db/dao/partyRecruitmentsDao');
 const { getInteractionData, getUserName, checkTimeRegex, setEmbed, setActionRow } = require('../../common/commandFunc');
+const errorHandler = require("../../common/errorHandler");
 
 module.exports = {
     data: new SlashCommandBuilder()
@@ -42,47 +43,49 @@ module.exports = {
                 .setDescription('게임의 이름, 롤체 티어 등을 입력')
         ),
     async execute(interaction) {
-        // 상호작용 데이터 가져오기
-        const interactionData = await getInteractionData(interaction, 'etc');
-        const {gameMode, startTime, minMembers, maxMembers} = interactionData;
-        // 사용자 닉네임 가져오기
-        const lolName = await getUserName(interaction);
+        await errorHandler(interaction, async (interaction) => {
+            // 상호작용 데이터 가져오기
+            const interactionData = await getInteractionData(interaction, 'etc');
+            const {gameMode, startTime, minMembers, maxMembers} = interactionData;
+            // 사용자 닉네임 가져오기
+            const lolName = await getUserName(interaction);
 
-        // 시간 정규식 확인
-        if (!checkTimeRegex(startTime)) {
-            await interaction.reply({
-                content: '⛔ 유효한 시간 형식이 아닙니다! 00:00 ~ 23:59 형식으로 입력하세요.',
-                ephemeral: true
+            // 시간 정규식 확인
+            if (!checkTimeRegex(startTime)) {
+                await interaction.reply({
+                    content: '⛔ 유효한 시간 형식이 아닙니다! 00:00 ~ 23:59 형식으로 입력하세요.',
+                    ephemeral: true
+                });
+                return;
+            }
+
+            // 임베드 생성
+            const embed = await setEmbed(interaction, interactionData, lolName, 'etc');
+
+            // 버튼 생성
+            const actionRow = await setActionRow('join');
+
+            // 메시지 전송
+            const message = await interaction.reply({
+                content: `@everyone (1/${maxMembers}) ${lolName}님의 ${gameMode} 구인이 시작되었어요!`,
+                embeds: [embed],
+                components: [actionRow],
+                allowedMentions: { parse: ['everyone'] },
+                fetchReply: true
             });
-            return;
-        }
 
-        // 임베드 생성
-        const embed = await setEmbed(interaction, interactionData, lolName, 'etc');
-
-        // 버튼 생성
-        const actionRow = await setActionRow('join');
-
-        // 메시지 전송
-        const message = await interaction.reply({
-            content: `@everyone (1/${maxMembers}) ${lolName}님의 ${gameMode} 구인이 시작되었어요!`,
-            embeds: [embed],
-            components: [actionRow],
-            allowedMentions: { parse: ['everyone'] },
-            fetchReply: true
-        });
-
-        const data = {
-            interactionId: message.interaction.id,
-            messageId : message.id,
-            owner: { id: interaction.user.id, name: lolName },
-            members: [{ id: interaction.user.id }],
-            maxMembers,
-            minMembers,
-            startTime,
-            channelId: process.env.LFP_ETC_GAME,
-            gameMode
-        };
-        await partyRecruitmentsDao.savePartyRecruitment(data);
+            const data = {
+                interactionId: message.interaction.id,
+                messageId : message.id,
+                owner: { id: interaction.user.id, name: lolName },
+                members: [{ id: interaction.user.id }],
+                maxMembers,
+                minMembers,
+                startTime,
+                channelId: process.env.LFP_ETC_GAME,
+                gameMode
+            };
+            await partyRecruitmentsDao.savePartyRecruitment(data);
+        })
     }
 };
