+++ commands/utility/nomalGame.js
@@ -1,6 +1,7 @@
 const { SlashCommandBuilder } = require('discord.js');
 const partyRecruitmentsDao = require('../../db/dao/partyRecruitmentsDao');
 const { getInteractionData, getUserName, checkTimeRegex, setEmbed, setActionRow } = require('../../common/commandFunc');
+const errorHandler = require("../../common/errorHandler");
 
 require('dotenv').config();
 
@@ -60,54 +61,56 @@ module.exports = {
                 .setDescription('본인 외의 추가 인원 선택')
         ),
     async execute(interaction) {
-        // 상호작용 데이터 가져오기
-        const interactionData = await getInteractionData(interaction);
-        const {gameMode, startTime, minMembers, extraMembers} = interactionData;
+        await errorHandler(interaction, async (interaction) => {
+            // 상호작용 데이터 가져오기
+            const interactionData = await getInteractionData(interaction);
+            const {gameMode, startTime, minMembers, extraMembers} = interactionData;
 
-        // 사용자 닉네임 가져오기
-        const lolName = await getUserName(interaction);
+            // 사용자 닉네임 가져오기
+            const lolName = await getUserName(interaction);
 
-        // 시간 정규식 확인
-        if (!checkTimeRegex(startTime)) {
-            await interaction.reply({
-                content: '⛔ 유효한 시간 형식이 아닙니다! 00:00 ~ 23:59 형식으로 입력하세요.',
-                ephemeral: true
+            // 시간 정규식 확인
+            if (!checkTimeRegex(startTime)) {
+                await interaction.reply({
+                    content: '⛔ 유효한 시간 형식이 아닙니다! 00:00 ~ 23:59 형식으로 입력하세요.',
+                    ephemeral: true
+                });
+                return;
+            }
+
+            // 임베드 생성
+            const embed = await setEmbed(interaction, interactionData, lolName);
+
+            // 버튼 생성
+            const actionRow = await setActionRow('join');
+
+            const currentMembers = extraMembers.length + 1;
+
+            // 메시지 전송
+            const message = await interaction.reply({
+                content: `@everyone (${currentMembers}/5)${lolName}님의 ${gameMode} 구인이 시작되었어요!`,
+                embeds: [embed],
+                components: [actionRow],
+                allowedMentions: { parse: ['everyone'] },
+                fetchReply: true
             });
-            return;
-        }
 
-        // 임베드 생성
-        const embed = await setEmbed(interaction, interactionData, lolName);
+            // 데이터베이스 구인글 저장
+            const data = {
+                interactionId: message.interaction.id,
+                messageId: message.id,
+                owner: { id: interaction.user.id, name: lolName },
+                members: [
+                    ...extraMembers.map(memberId => ({ id: memberId, message: null })) // 추가 인원 변환
+                ],
+                minMembers,
+                currentMembers,
+                startTime,
+                channelId: process.env.LFP_NORMAL_GAME,
+                gameMode
+            };
 
-        // 버튼 생성
-        const actionRow = await setActionRow('join');
-
-        const currentMembers = extraMembers.length + 1;
-
-        // 메시지 전송
-        const message = await interaction.reply({
-            content: `@everyone (${currentMembers}/5)${lolName}님의 ${gameMode} 구인이 시작되었어요!`,
-            embeds: [embed],
-            components: [actionRow],
-            allowedMentions: { parse: ['everyone'] },
-            fetchReply: true
-        });
-
-        // 데이터베이스 구인글 저장
-        const data = {
-            interactionId: message.interaction.id,
-            messageId: message.id,
-            owner: { id: interaction.user.id, name: lolName },
-            members: [
-                ...extraMembers.map(memberId => ({ id: memberId, message: null })) // 추가 인원 변환
-            ],
-            minMembers,
-            currentMembers,
-            startTime,
-            channelId: process.env.LFP_NORMAL_GAME,
-            gameMode
-        };
-
-        await partyRecruitmentsDao.savePartyRecruitment(data);
+            await partyRecruitmentsDao.savePartyRecruitment(data);
+        })
     }
 };
