+++ events/modalInteraction.js
